% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miregression-package.R
\docType{package}
\name{miregression-package}
\alias{miregression}
\alias{miregression-package}
\title{miregression: Multiple imputation regression}
\description{
A package to run regression models using multiple imputation through the Amelia package
}
\keyword{internal}
