% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelia_mi_model.R
\name{stargazer_amelia_mi_model}
\alias{stargazer_amelia_mi_model}
\title{Use stargazer to make output}
\usage{
stargazer_amelia_mi_model(
  object,
  digits = 3,
  include_missingess_diagnostics = F,
  remove_rows = NULL,
  ...
)
}
\arguments{
\item{object}{an amelia_mi_model object}

\item{digits}{number of digits to round to}

\item{include_missingess_diagnostics}{Include missingness diagnostics,  df, r, and miss.info, in the table}

\item{remove_rows}{Rows to remove from the output (eg fixed effects)}

\item{...}{Other arguments to be passed to stargazer}
}
\value{
Stargazer table of the model object
}
\description{
Use stargazer to make output
}
\examples{
stargazer_amelia_mi_model(m1,type='text',out='table.txt')
}
